requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-extra-types>=2.0.0
pydantic-settings>=2.0.0
pydantic_core>=2.0.0
fastapi>=0.100.0
faststream>=0.2.0
aiokafka>=0.8.0
uvicorn>=0.22.0
python-json-logger>=2.0.7
celery>=5.3.0
redis>=4.6.0
sse-starlette>=1.6.0
prometheus-fastapi-instrumentator>=6.0.0
saq[web,redis]>=0.9.0
httpx>=0.24.0

# Agent dependencies
together>=0.1.5
browser-use>=0.1.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Browser automation
playwright>=1.35.0

# Utilities
uuid>=1.30.0
python-multipart>=0.0.6